feat: PRO-2373 refactor the example scripts by grouping
