--- examples/15-disable-sessionkey-module copy.ts
@@ -1,52 +1,0 @@
-import { EtherspotBundler, ModularSdk, SessionKeyValidator } from '../src';
-import * as dotenv from 'dotenv';
-import { getViemAccount, sleep } from '../src/sdk/common';
-
-dotenv.config();
-
-async function main() {
-  const bundlerApiKey = 'eyJvcmciOiI2NTIzZjY5MzUwOTBmNzAwMDFiYjJkZWIiLCJpZCI6IjMxMDZiOGY2NTRhZTRhZTM4MGVjYjJiN2Q2NDMzMjM4IiwiaCI6Im11cm11cjEyOCJ9';
-
-  // initializating sdk...
-  const modularSdk = new ModularSdk(
-    getViemAccount(process.env.WALLET_PRIVATE_KEY),
-    {
-      chainId: Number(process.env.CHAIN_ID),
-      bundlerProvider: new EtherspotBundler(Number(process.env.CHAIN_ID), bundlerApiKey)
-    })
-
-  console.log('address: ', modularSdk.getEOAAddress());
-
-  // get address of EtherspotWallet
-  const address: string = await modularSdk.getCounterFactualAddress();
-
-  console.log('\x1b[33m%s\x1b[0m', `EtherspotWallet address: ${address}`);
-
-  // get instance  of SessionKeyValidator
-  const sessionKeyModule = new SessionKeyValidator(
-    modularSdk,
-    new EtherspotBundler(Number(process.env.CHAIN_ID), bundlerApiKey)
-  )
-
-  const response = await sessionKeyModule.disableSessionKey();
-
-  console.log('\x1b[33m%s\x1b[0m', `UserOpHash: `, response.userOpHash);
-  console.log('\x1b[33m%s\x1b[0m', `SessionKey: `, response.sessionKey);
-
-  // get transaction hash...
-  console.log('Waiting for transaction...');
-  let userOpsReceipt = null;
-  const timeout = Date.now() + 60000; // 1 minute timeout
-  while ((userOpsReceipt == null) && (Date.now() < timeout)) {
-    await sleep(2);
-    userOpsReceipt = await modularSdk.getUserOpReceipt(response.userOpHash);
-  }
-  console.log('\x1b[33m%s\x1b[0m', `Transaction Receipt: `, userOpsReceipt);
-
-  const sessionKeys = await sessionKeyModule.getAssociatedSessionKeys();
-  console.log('\x1b[33m%s\x1b[0m', `AssociatedSessionKeys: `, sessionKeys);
-}
-
-main()
-  .catch(console.error)
-  .finally(() => process.exit());
