+++ examples/basics/add-guardians.ts
@@ -1,13 +1,12 @@
-import { EtherspotBundler, ModularSdk } from '../src';
-import { printOp } from '../src/sdk/common/OperationUtils';
+import { printOp } from '../../src/sdk/common/OperationUtils';
 import * as dotenv from 'dotenv';
-import { sleep } from '../src/sdk/common';
-import { getViemAccount } from '../src/sdk/common/utils/viem-utils';
+import { sleep } from '../../src/sdk/common';
 import { encodeFunctionData, parseAbi } from 'viem';
-import { generateModularSDKInstance } from './helpers/sdk-helper';
+import { generateModularSDKInstance } from '../helpers/sdk-helper';
 
 dotenv.config();
 
+// tsx examples/basics/add-guardians.ts
 async function main() {
   const bundlerApiKey = 'eyJvcmciOiI2NTIzZjY5MzUwOTBmNzAwMDFiYjJkZWIiLCJpZCI6IjMxMDZiOGY2NTRhZTRhZTM4MGVjYjJiN2Q2NDMzMjM4IiwiaCI6Im11cm11cjEyOCJ9';
