+++ examples/basics/callGasLimit.ts
@@ -1,18 +1,15 @@
-import { EtherspotBundler, ModularSdk } from '../src';
-import { printOp } from '../src/sdk/common/OperationUtils';
+import { printOp } from '../../src/sdk/common/OperationUtils';
 import * as dotenv from 'dotenv';
-import { sleep } from '../src/sdk/common';
-import { getViemAccount } from '../src/sdk/common/utils/viem-utils';
+import { sleep } from '../../src/sdk/common';
 import { parseEther } from 'viem';
-import { generateModularSDKInstance } from './helpers/sdk-helper';
+import { generateModularSDKInstance } from '../helpers/sdk-helper';
 
 dotenv.config();
 
 const recipient = '0x80a1874E1046B1cc5deFdf4D3153838B72fF94Ac'; // recipient wallet address
-const value = '0.0000001'; // transfer value
-const apiKey = 'arka_public_key'; // Only testnets are available, if you need further assistance in setting up a paymaster service for your dapp, please reach out to us on discord or https://etherspot.fyi/arka/intro
+const value = '0.0001'; // transfer value
 const bundlerApiKey = 'eyJvcmciOiI2NTIzZjY5MzUwOTBmNzAwMDFiYjJkZWIiLCJpZCI6IjMxMDZiOGY2NTRhZTRhZTM4MGVjYjJiN2Q2NDMzMjM4IiwiaCI6Im11cm11cjEyOCJ9';
-// tsx 
+
 async function main() {
   // initializating sdk...
   const modularSdk = generateModularSDKInstance(
@@ -41,9 +38,8 @@ async function main() {
   console.log('balances: ', balance);
 
   // estimate transactions added to the batch and get the fee data for the UserOp
-  const op = await modularSdk.estimate({
-    paymasterDetails: { url: `https://arka.etherspot.io?apiKey=${apiKey}&chainId=${Number(process.env.CHAIN_ID)}`, context: { mode: 'sponsor' } }
-  });
+  // passing callGasLimit as 40000 to manually set it
+  const op = await modularSdk.estimate({ callGasLimit: 4000 });
   console.log(`Estimate UserOp: ${await printOp(op)}`);
 
   // sign the UserOp and sending to the bundler...
