+++ examples/basics/concurrent-userops.ts
@@ -1,10 +1,10 @@
-import { printOp } from '../src/sdk/common/OperationUtils';
+import { printOp } from '../../src/sdk/common/OperationUtils';
 import * as dotenv from 'dotenv';
-import { sleep } from '../src/sdk/common';
-import { getPublicClient } from '../src/sdk/common/utils/viem-utils';
+import { sleep } from '../../src/sdk/common';
+import { getPublicClient } from '../../src/sdk/common/utils/viem-utils';
 import { Hex, http, parseEther, PublicClient } from 'viem';
-import { generateModularSDKInstance } from './helpers/sdk-helper';
-import { BigNumber } from '../src/sdk/types/bignumber';
+import { generateModularSDKInstance } from '../helpers/sdk-helper';
+import { BigNumber } from '../../src/sdk/types/bignumber';
 
 dotenv.config();
