+++ examples/basics/transfer-funds.ts
@@ -1,15 +1,16 @@
-import { printOp } from '../src/sdk/common/OperationUtils';
+import { printOp } from '../../src/sdk/common/OperationUtils';
 import * as dotenv from 'dotenv';
-import { sleep } from '../src/sdk/common';
+import { sleep } from '../../src/sdk/common';
 import { parseEther } from 'viem';
-import { generateModularSDKInstance } from './helpers/sdk-helper';
+import { generateModularSDKInstance } from '../helpers/sdk-helper';
 
 dotenv.config();
 
 const recipient = '0x80a1874E1046B1cc5deFdf4D3153838B72fF94Ac'; // recipient wallet address
-const value = '0.0001'; // transfer value
+const value = '0.0000001'; // transfer value
 const bundlerApiKey = 'eyJvcmciOiI2NTIzZjY5MzUwOTBmNzAwMDFiYjJkZWIiLCJpZCI6IjMxMDZiOGY2NTRhZTRhZTM4MGVjYjJiN2Q2NDMzMjM4IiwiaCI6Im11cm11cjEyOCJ9';
 
+// tsx examples/basics/transfer-funds.ts
 async function main() {
   // initializating sdk...
   const modularSdk = generateModularSDKInstance(
@@ -17,7 +18,6 @@ async function main() {
     Number(process.env.CHAIN_ID),
     bundlerApiKey
   );// Testnets dont need apiKey on bundlerProvider
-
 
   console.log('address: ', modularSdk.getEOAAddress());
 
@@ -38,8 +38,7 @@ async function main() {
   console.log('balances: ', balance);
 
   // estimate transactions added to the batch and get the fee data for the UserOp
-  // passing callGasLimit as 40000 to manually set it
-  const op = await modularSdk.estimate({ callGasLimit: 4000 });
+  const op = await modularSdk.estimate();
   console.log(`Estimate UserOp: ${await printOp(op)}`);
 
   // sign the UserOp and sending to the bundler...
@@ -49,7 +48,7 @@ async function main() {
   // get transaction hash...
   console.log('Waiting for transaction...');
   let userOpsReceipt = null;
-  const timeout = Date.now() + 60000; // 1 minute timeout
+  const timeout = Date.now() + 1200000; // 1 minute timeout
   while ((userOpsReceipt == null) && (Date.now() < timeout)) {
     await sleep(2);
     userOpsReceipt = await modularSdk.getUserOpReceipt(uoHash);
