+++ examples/modules/get-previous-address.ts
@@ -1,10 +1,11 @@
-import { EtherspotBundler, ModularSdk } from '../src';
+import { EtherspotBundler, ModularSdk } from '../../src';
 import * as dotenv from 'dotenv';
-import { getViemAccount } from '../src/sdk/common/utils';
+import { MODULE_TYPE } from '../../src/sdk/common';
+import { getViemAccount } from '../../src/sdk/common/utils/viem-utils';
 
 dotenv.config();
 
-// tsx examples/13-list-modules.ts
+// tsx examples/modules/get-previous-address.ts
 async function main() {
   const bundlerApiKey = 'eyJvcmciOiI2NTIzZjY5MzUwOTBmNzAwMDFiYjJkZWIiLCJpZCI6IjMxMDZiOGY2NTRhZTRhZTM4MGVjYjJiN2Q2NDMzMjM4IiwiaCI6Im11cm11cjEyOCJ9';
 
@@ -23,8 +24,8 @@ async function main() {
 
   console.log('\x1b[33m%s\x1b[0m', `EtherspotWallet address: ${address}`);
 
-  const moduleInfo = await modularSdk.getAllModules();
-  console.log(`moduleInfo: ${JSON.stringify(moduleInfo)}`);
+  const previousAddress = await modularSdk.getPreviousAddress(MODULE_TYPE.VALIDATOR, '0xFE14F6d4e407850b24D160B9ACfBb042D32BE492');
+  console.log(`previousAddress: ${previousAddress}`);
 }
 
 main()
