+++ examples/modules/list-modules.ts
@@ -1,11 +1,10 @@
-import { EtherspotBundler, ModularSdk } from '../src';
+import { EtherspotBundler, ModularSdk } from '../../src';
 import * as dotenv from 'dotenv';
-import { MODULE_TYPE, sleep } from '../src/sdk/common';
-import { getViemAccount } from '../src/sdk/common/utils/viem-utils';
+import { getViemAccount } from '../../src/sdk/common/utils';
 
 dotenv.config();
 
-// npx ts-node examples/12-get-previous-address.ts
+// tsx examples/modules/list-modules.ts
 async function main() {
   const bundlerApiKey = 'eyJvcmciOiI2NTIzZjY5MzUwOTBmNzAwMDFiYjJkZWIiLCJpZCI6IjMxMDZiOGY2NTRhZTRhZTM4MGVjYjJiN2Q2NDMzMjM4IiwiaCI6Im11cm11cjEyOCJ9';
 
@@ -24,8 +23,8 @@ async function main() {
 
   console.log('\x1b[33m%s\x1b[0m', `EtherspotWallet address: ${address}`);
 
-  const previousAddress = await modularSdk.getPreviousAddress(MODULE_TYPE.VALIDATOR, '0x409474DFf53dD5824a7bf62D183639263853e3D5');
-  console.log(`previousAddress: ${previousAddress}`);
+  const moduleInfo = await modularSdk.getAllModules();
+  console.log(`moduleInfo: ${JSON.stringify(moduleInfo)}`);
 }
 
 main()
