+++ examples/sessionkeys/disable-sessionkey-module.ts
@@ -1,16 +1,16 @@
-import { EtherspotBundler, ModularSdk, SessionKeyValidator } from '../src';
+import { EtherspotBundler, ModularSdk, SessionKeyValidator } from '../../src';
 import * as dotenv from 'dotenv';
-import { getViemAccount, sleep } from '../src/sdk/common';
-import { KeyStore } from '../src/sdk/SessionKeyValidator';
+import { getViemAccount, sleep } from '../../src/sdk/common';
 
 dotenv.config();
 
+// tsx examples/sessionkeys/disable-sessionkey-module.ts
 async function main() {
   const bundlerApiKey = 'eyJvcmciOiI2NTIzZjY5MzUwOTBmNzAwMDFiYjJkZWIiLCJpZCI6IjMxMDZiOGY2NTRhZTRhZTM4MGVjYjJiN2Q2NDMzMjM4IiwiaCI6Im11cm11cjEyOCJ9';
 
   // initializating sdk...
   const modularSdk = new ModularSdk(
-    getViemAccount(process.env.WALLET_PRIVATE_KEY),
+    getViemAccount(process.env.WALLET_PRIVATE_KEY as string),
     {
       chainId: Number(process.env.CHAIN_ID),
       bundlerProvider: new EtherspotBundler(Number(process.env.CHAIN_ID), bundlerApiKey)
@@ -23,26 +23,13 @@ async function main() {
 
   console.log('\x1b[33m%s\x1b[0m', `EtherspotWallet address: ${address}`);
 
-  const token = '0x1c7D4B196Cb0C7B01d743Fbc6116a902379C7238';
-  const functionSelector = '0xa9059cbb';
-  const spendingLimit = '100000';
-  const validAfter = new Date().getTime();
-  const validUntil = new Date().getTime() + 24 * 60 * 60 * 1000;
-
   // get instance  of SessionKeyValidator
   const sessionKeyModule = new SessionKeyValidator(
     modularSdk,
     new EtherspotBundler(Number(process.env.CHAIN_ID), bundlerApiKey)
   )
 
-  const response = await sessionKeyModule.enableSessionKey(
-    token,
-    functionSelector,
-    spendingLimit,
-    validAfter,
-    validUntil,
-    KeyStore.AWS
-  );
+  const response = await sessionKeyModule.disableSessionKey();
 
   console.log('\x1b[33m%s\x1b[0m', `UserOpHash: `, response.userOpHash);
   console.log('\x1b[33m%s\x1b[0m', `SessionKey: `, response.sessionKey);
