+++ examples/sessionkeys/rotate-sessionkey-module.ts
@@ -1,16 +1,17 @@
-import { EtherspotBundler, ModularSdk, SessionKeyValidator } from '../src';
+import { EtherspotBundler, ModularSdk, SessionKeyValidator } from '../../src';
 import * as dotenv from 'dotenv';
-import { getViemAccount, sleep } from '../src/sdk/common';
-import { KeyStore } from '../src/sdk/SessionKeyValidator';
+import { getViemAccount, sleep } from '../../src/sdk/common';
+import { KeyStore } from '../../src/sdk/SessionKeyValidator';
 
 dotenv.config();
 
+// tsx examples/sessionkeys/rotate-sessionkey-module.ts
 async function main() {
   const bundlerApiKey = 'eyJvcmciOiI2NTIzZjY5MzUwOTBmNzAwMDFiYjJkZWIiLCJpZCI6IjMxMDZiOGY2NTRhZTRhZTM4MGVjYjJiN2Q2NDMzMjM4IiwiaCI6Im11cm11cjEyOCJ9';
 
   // initializating sdk...
   const modularSdk = new ModularSdk(
-    getViemAccount(process.env.WALLET_PRIVATE_KEY),
+    getViemAccount(process.env.WALLET_PRIVATE_KEY as string),
     {
       chainId: Number(process.env.CHAIN_ID),
       bundlerProvider: new EtherspotBundler(Number(process.env.CHAIN_ID), bundlerApiKey)
