+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@etherspot/modular-sdk",
-  "version": "1.1.4",
+  "version": "1.2.0",
   "description": "Etherspot Modular SDK - build with ERC-7579 smart accounts modules",
   "keywords": [
     "ether",
@@ -25,19 +25,20 @@
   "types": "./dist/index.d.ts",
   "scripts": {
     "start": "exit 1",
-    "01-get-address": "./node_modules/.bin/ts-node ./examples/01-get-address",
-    "02-transfer-funds": "./node_modules/.bin/ts-node ./examples/02-transfer-funds",
-    "03-transfer-erc20": "./node_modules/.bin/ts-node ./examples/03-transfer-erc20",
-    "04-transfer-nft": "./node_modules/.bin/ts-node ./examples/04-transfer-nft",
-    "05-add-guardians": "./node_modules/.bin/ts-node ./examples/05-add-guardians",
-    "06-paymaster": "./node_modules/.bin/ts-node ./examples/06-paymaster",
-    "07-paymaster-arka": "./node_modules/.bin/ts-node ./examples/07-paymaster-arka",
-    "08-paymaster-validUntil-validAfter": "./node_modules/.bin/ts-node ./examples/08-paymaster-validUntil-validAfter",
-    "09-callGasLimit": "./node_modules/.bin/ts-node ./examples/09-callGasLimit",
-    "10-get-multiple-accounts": "./node_modules/.bin/ts-node ./examples/10-get-multiple-accounts",
-    "11-concurrent-userops": "./node_modules/.bin/ts-node ./examples/11-concurrent-userops",
-    "14-install-module": "./node_modules/.bin/ts-node ./examples/14-install-module",
-    "15-uninstall-module": "./node_modules/.bin/ts-node ./examples/15-uninstall-module",
+    "get-address": "tsx ./examples/basics/get-address",
+    "transfer-funds": "tsx ./examples/basics/transfer-funds",
+    "transfer-erc20": "tsx ./examples/basics/transfer-erc20",
+    "transfer-nft": "tsx ./examples/basics/transfer-nft",
+    "add-guardians": "tsx ./examples/basics/add-guardians",
+    "paymaster": "tsx ./examples/paymaster/paymaster",
+    "callGasLimit": "tsx ./examples/basics/callGasLimit",
+    "get-multiple-accounts": "tsx ./examples/basics/get-multiple-accounts",
+    "concurrent-userops": "tsx ./examples/basics/concurrent-userops",
+    "install-module": "tsx ./examples/modules/install-module",
+    "uninstall-module": "tsx ./examples/modules/uninstall-module",
+    "enable-sessionkey-module": "tsx ./examples/sessionkeys/enable-sessionkey-module",
+    "rotate-sessionkey-module": "tsx ./examples/sessionkeys/rotate-sessionkey-module",
+    "disable-sessionkey-module": "tsx ./examples/sessionkeys/disable-sessionkey-module",
     "format": "prettier --write \"{src,test,examples}/**/*.ts\"",
     "lint": "eslint \"{src,test,examples}/**/*.ts\"",
     "lint-fix": "npm run lint -- --fix",
